+++ src/App.jsx
@@ -9,7 +9,9 @@ function App() {
 
   return (
     <>
+    <div>
     <Navigation/>
+    </div>
       <div>
         <a href='https://vitejs.dev' target='_blank'>
           <img src={viteLogo} className='logo' alt='Vite logo' />
